# .futurxlab/api-specification.yaml
openapi: 3.0.3
info:
  title: InKnowing API
  description: |
    InKnowing is an AI-powered book dialogue platform that enables users to have conversations with books and book characters.

    ## Business Logic Conservation
    Based on the principle of business logic conservation, this API documentation maintains complete mapping between:
    - User Journey → API Endpoints
    - State Transitions → API Operations
    - Sequence Interactions → API Calls
    - Business Rules → API Responses

    ## Key Features
    - Question-driven and book-driven discovery
    - AI-powered dialogues with books and characters
    - User book upload with vectorization
    - Multi-tier membership system
    - Admin management dashboard

    ## Authentication
    The API uses JWT Bearer tokens for authentication. Tokens are obtained through the login endpoints and must be included in the Authorization header for protected endpoints.

    ## Rate Limiting
    - Free users: 20 requests per day
    - Basic members: 200 requests per month
    - Premium members: 500 requests per month
    - Super members: 1000 requests per month
  version: 1.0.0
  contact:
    name: InKnowing Support
    email: support@inknowing.ai
    url: https://inknowing.ai/support
  license:
    name: Proprietary
    url: https://inknowing.ai/terms
servers:
  - url: https://api.inknowing.ai/v1
    description: Production server
  - url: https://staging-api.inknowing.ai/v1
    description: Staging server
...
paths:
...
  /dialogues/book/start:
    post:
      tags:
        - Dialogues
      summary: Start book dialogue
      description: |
        Initialize a dialogue session with a book
        Business Logic: User Action → System Response Sequence
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - book_id
              properties:
                book_id:
                  type: string
                initial_question:
                  type: string
      responses:
        '201':
          description: Dialogue started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DialogueSession'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Quota exceeded
        '404':
          $ref: '#/components/responses/NotFound'
  /dialogues/character/start:
    post:
      tags:
        - Dialogues
      summary: Start character dialogue
      description: |
        Initialize a dialogue session with a book character
        Business Logic: Character Selection → Immersive Dialogue State
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - book_id
                - character_id
              properties:
                book_id:
                  type: string
                character_id:
                  type: string
                initial_message:
                  type: string
      responses:
        '201':
          description: Character dialogue started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DialogueSession'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Quota exceeded
        '404':
          $ref: '#/components/responses/NotFound'
  /dialogues/{sessionId}/messages:
    post:
      tags:
        - Dialogues
      summary: Send message
      description: Send a message in an active dialogue session
      security:
        - bearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - message
              properties:
                message:
                  type: string
                  minLength: 1
                  maxLength: 2000
      responses:
        '200':
          description: Message sent and response received
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DialogueMessage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Quota exceeded
        '404':
          $ref: '#/components/responses/NotFound'

    get:
      tags:
        - Dialogues
      summary: Get dialogue history
      description: Get message history for a dialogue session
      security:
        - bearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Message history
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/DialogueMessage'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /dialogues/{sessionId}/context:
    get:
      tags:
        - Dialogues
      summary: Get dialogue context
      description: Get current context and references for dialogue session
      security:
        - bearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Dialogue context
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DialogueContext'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /dialogues/history:
    get:
      tags:
        - Dialogues
      summary: Get user dialogue history
      description: Get all dialogue sessions for current user
      security:
        - bearerAuth: []
      parameters:
        - name: book_id
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [book, character]
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        '200':
          description: Dialogue history
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessions:
                    type: array
                    items:
                      $ref: '#/components/schemas/DialogueSession'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ==================== Uploads ====================
...
components:
  schemas:
...
    DialogueSession:
      type: object
      properties:
        id:
          type: string
        book_id:
          type: string
        book_title:
          type: string
        type:
          type: string
          enum: [book, character]
        character_id:
          type: string
          nullable: true
        character_name:
          type: string
          nullable: true
        user_id:
          type: string
        message_count:
          type: integer
        last_message_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        status:
          type: string
          enum: [active, ended]
    DialogueMessage:
      type: object
      properties:
        id:
          type: string
        session_id:
          type: string
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
        references:
          type: array
          items:
            $ref: '#/components/schemas/Reference'
        timestamp:
          type: string
          format: date-time
        tokens_used:
          type: integer
        model_used:
          type: string
    DialogueContext:
      type: object
      properties:
        session_id:
          type: string
        book_context:
          type: object
          properties:
            current_chapter:
              type: string
              nullable: true
            discussed_topics:
              type: array
              items:
                type: string
            key_references:
              type: array
              items:
                $ref: '#/components/schemas/Reference'
        character_context:
          type: object
          nullable: true
          properties:
            character_state:
              type: string
            emotional_tone:
              type: string
            remembered_facts:
              type: array
              items:
                type: string
    Reference:
      type: object
      properties:
        type:
          type: string
          enum: [chapter, page, paragraph, character_memory]
        chapter:
          type: integer
          nullable: true
        page:
          type: integer
          nullable: true
        text:
          type: string
        highlight:
          type: string

    # ==================== Upload Schemas ====================
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
        has_next:
          type: boolean
        has_prev:
          type: boolean
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true
        timestamp:
          type: string
          format: date-time
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for user authentication

    adminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for admin authentication

# API Response Standards
...